Alarm.h - Some fixes to make
Currently writes to file and can open and parse contents of file. removeAlarm() has some issues currently with properly removing an alarm. Contents of the file also need to be parsed into vector<Node> alarms so that the alarms can be sorted, then replaced into file (e.g. Schedule.txt) so that the file can be truncated and replaced with the sorted contents of alarms.

Use of vector may be replaced with std::list to ensure that sorting of the alarms is easier.